This series adds basic support for the vector crypto extensions.
Specifically the new extensions are:
- Zvbb (bit manipulation)
- Zvbc (carryless multiplication)
- Zvkg (gcm acceleration)
- Zvkned (nist encryption/decryption)
- Zvknh[a,b] (nist hash algorithms)
- Zvkn (nist suite)
- Zvkng (nist suite with gcm)
- Zvknc (nist suite with carryless multiplicatoin)
- Zvksed (shang-mi encryption/decryption)
- Zvksh (shang-mi hash algorithms)
- Zvks (shang-mi suite)
- Zvksg (shang-mi with gcm)
- Zvksc (shang-mi with carryless multiplication)

The implementation follows the version 20230531 of the specification,
which can be found here:
  https://github.com/riscv/riscv-crypto/releases/tag/v20230531

Note, that this specification is not frozen yet, meaning that
incompatible changes are possible.
Therefore, this patchset is marked as RFC and should not be considered
for upstream inclusion.

All extensions come with (passing) tests for the feature test macros.

A Binutils patch series for vector crypto support can be found here:
  https://github.com/nhukc/binutils-gdb/commits/master

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>
[Updated to most recent version of specification]
Signed-off-by: Nathan Huckleberry <nhuck@google.com>